Move clean up functions to their managers file
