+++ felix/dataplane/linux/int_dataplane.go
@@ -1729,94 +1729,6 @@ func ConfigureDefaultMTUs(hostMTU int, c *Config) {
 	}
 }
 
-func cleanUpIPIPAddrs() {
-	// If IPIP is not enabled, check to see if there is are addresses in the IPIP device and delete them if there are.
-	log.Debug("Checking if we need to clean up the IPIP device")
-
-	var errFound bool
-
-cleanupRetry:
-	for i := 0; i <= maxCleanupRetries; i++ {
-		errFound = false
-		if i > 0 {
-			log.Debugf("Retrying %v/%v times", i, maxCleanupRetries)
-		}
-		link, err := netlink.LinkByName(dataplanedefs.IPIPIfaceName)
-		if err != nil {
-			if _, ok := err.(netlink.LinkNotFoundError); ok {
-				log.Debug("IPIP disabled and no IPIP device found")
-				return
-			}
-			log.WithError(err).Warn("IPIP disabled and failed to query IPIP device.")
-			errFound = true
-
-			// Sleep for 1 second before retrying
-			time.Sleep(1 * time.Second)
-			continue
-		}
-		addrs, err := netlink.AddrList(link, netlink.FAMILY_V4)
-		if err != nil {
-			log.WithError(err).Warn("IPIP disabled and failed to list addresses, will be unable to remove any old addresses from the device should they exist.")
-			errFound = true
-
-			// Sleep for 1 second before retrying
-			time.Sleep(1 * time.Second)
-			continue
-		}
-
-		for _, oldAddr := range addrs {
-			if err := netlink.AddrDel(link, &oldAddr); err != nil {
-				log.WithError(err).Errorf("IPIP disabled and failed to delete unwanted IPIP address %s.", oldAddr.IPNet)
-				errFound = true
-
-				// Sleep for 1 second before retrying
-				time.Sleep(1 * time.Second)
-				continue cleanupRetry
-			}
-		}
-	}
-	if errFound {
-		log.Warnf("Giving up trying to clean up IPIP addresses after retrying %v times", maxCleanupRetries)
-	}
-}
-
-func cleanUpVXLANDevice(deviceName string) {
-	// If VXLAN is not enabled, check to see if there is a VXLAN device and delete it if there is.
-	log.Debug("Checking if we need to clean up the VXLAN device")
-
-	var errFound bool
-	for i := 0; i <= maxCleanupRetries; i++ {
-		errFound = false
-		if i > 0 {
-			log.Debugf("Retrying %v/%v times", i, maxCleanupRetries)
-		}
-		link, err := netlink.LinkByName(deviceName)
-		if err != nil {
-			if _, ok := err.(netlink.LinkNotFoundError); ok {
-				log.Debug("VXLAN disabled and no VXLAN device found")
-				return
-			}
-			log.WithError(err).Warn("VXLAN disabled and failed to query VXLAN device.")
-			errFound = true
-
-			// Sleep for 1 second before retrying
-			time.Sleep(1 * time.Second)
-			continue
-		}
-		if err = netlink.LinkDel(link); err != nil {
-			log.WithError(err).Error("VXLAN disabled and failed to delete unwanted VXLAN device.")
-			errFound = true
-
-			// Sleep for 1 second before retrying
-			time.Sleep(1 * time.Second)
-			continue
-		}
-	}
-	if errFound {
-		log.Warnf("Giving up trying to clean up VXLAN device after retrying %v times", maxCleanupRetries)
-	}
-}
-
 type Manager interface {
 	// OnUpdate is called for each protobuf message from the datastore.  May either directly
 	// send updates to the IPSets and generictables.Table objects (which will queue the updates
