+++ felix/dataplane/linux/ipip_mgr.go
@@ -22,6 +22,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
 
+	"github.com/projectcalico/calico/felix/dataplane/linux/dataplanedefs"
 	"github.com/projectcalico/calico/felix/ip"
 	"github.com/projectcalico/calico/felix/logutils"
 	"github.com/projectcalico/calico/felix/netlinkshim"
@@ -82,7 +83,6 @@ func newIPIPManagerWithShims(
 	opRecorder logutils.OpRecorder,
 	nlHandle netlinkshim.Interface,
 ) *ipipManager {
-
 	if ipVersion != 4 {
 		logrus.Errorf("IPIP manager only supports IPv4")
 		return nil
@@ -205,4 +205,55 @@ func (m *ipipManager) device(_ netlink.Link) (netlink.Link, string, error) {
 		return nil, "", fmt.Errorf("address is not set")
 	}
 	return ipip, address.String(), nil
+}
+
+func cleanUpIPIPAddrs() {
+	// If IPIP is not enabled, check to see if there is are addresses in the IPIP device and delete them if there are.
+	logrus.Debug("Checking if we need to clean up the IPIP device")
+
+	var errFound bool
+
+cleanupRetry:
+	for i := 0; i <= maxCleanupRetries; i++ {
+		errFound = false
+		if i > 0 {
+			logrus.Debugf("Retrying %v/%v times", i, maxCleanupRetries)
+		}
+		link, err := netlink.LinkByName(dataplanedefs.IPIPIfaceName)
+		if err != nil {
+			if _, ok := err.(netlink.LinkNotFoundError); ok {
+				logrus.Debug("IPIP disabled and no IPIP device found")
+				return
+			}
+			logrus.WithError(err).Warn("IPIP disabled and failed to query IPIP device.")
+			errFound = true
+
+			// Sleep for 1 second before retrying
+			time.Sleep(1 * time.Second)
+			continue
+		}
+		addrs, err := netlink.AddrList(link, netlink.FAMILY_V4)
+		if err != nil {
+			logrus.WithError(err).Warn("IPIP disabled and failed to list addresses, will be unable to remove any old addresses from the device should they exist.")
+			errFound = true
+
+			// Sleep for 1 second before retrying
+			time.Sleep(1 * time.Second)
+			continue
+		}
+
+		for _, oldAddr := range addrs {
+			if err := netlink.AddrDel(link, &oldAddr); err != nil {
+				logrus.WithError(err).Errorf("IPIP disabled and failed to delete unwanted IPIP address %s.", oldAddr.IPNet)
+				errFound = true
+
+				// Sleep for 1 second before retrying
+				time.Sleep(1 * time.Second)
+				continue cleanupRetry
+			}
+		}
+	}
+	if errFound {
+		logrus.Warnf("Giving up trying to clean up IPIP addresses after retrying %v times", maxCleanupRetries)
+	}
 }
